Refactor code into directory tree
