+++ .vscode/settings.json
@@ -1,0 +1,14 @@
+// Place your settings in this file to overwrite default and user settings.
+// See http://stackoverflow.com/a/32604188/266535
+{
+    "typescript.tsdk": "./node_modules/typescript/lib",
+    "files.exclude": {
+        "**/.git": true,
+        "**/.DS_Store": true,
+        "**/*.map": true,
+        "**/*.js": { "when": "$(basename).ts"},
+        "**/*?.js": { "when": "$(basename).tsx"},
+        "node_modules": true,
+        "dist": true
+    }
+}
