+++ src/shared/person.d.ts
