+++ src/views/pages/everyone.ejs
